:recycle: refactor(service) : test code 위치 변경 ( lottoStore -> lotto 객체로 메소드 위치 변경)
