+++ src/main/java/lotto/model/Lotto.java
@@ -1,8 +1,6 @@
-package lotto;
+package lotto.model;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Random;
 
 public class Lotto {
