+++ src/test/java/lotto/model/LottoTest.java
@@ -1,0 +1,16 @@
+package lotto.model;
+
+import org.junit.jupiter.api.Test;
+
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+class LottoTest {
+
+    @Test
+    void shouldReturnLottoMatchedCount(){
+        Lotto lotto = new Lotto(List.of(1, 4, 5, 6, 9));
+        assertThat(lotto.getMatchedCount(List.of(1, 4, 5, 6, 3))).isEqualTo(4L);
+    }
+}
